code style fixes CR-1047439 multi_process example fails for zcu102_base_dfx hardware board run

+++ src/runtime_src/core/edge/drm/zocl/zocl_cu.c
@@ -105,7 +105,7 @@ zocl_cu_status_print(struct zocl_cu *cu)
 	struct zcu_core *cu_core = cu->core;
 
 	DRM_INFO("addr 0x%llx, status 0x%x",
-	    cu_core->paddr, ioread32(cu_core->vaddr));
+	    (u64)cu_core->paddr, ioread32(cu_core->vaddr));
 }
 
 /* -- HLS adapter start -- */
